Move views into their own files and create a folder hierarchy
