+++ LumiformApp/LumiformApp/Form View/Views/FormView.swift
@@ -1,5 +1,5 @@
 //
-//  ContentView.swift
+//  FormView.swift
 //  LumiformApp
 //
 //  Created by Aly Yakan on 25/07/2025.
@@ -8,29 +8,7 @@
 import SwiftUI
 import Lumiform
 
-class FormViewModel: ObservableObject {
-    @Published var rootItem: FormItem?
-    @Published var error: String?
-
-    private let loader: FormLoader
-
-    init(loader: FormLoader) {
-        self.loader = loader
-    }
-
-    func loadData() {
-        loader.load { [weak self] result in
-            switch result {
-            case .success(let form):
-                self?.rootItem = form.rootPage
-            case .failure:
-                self?.error = "Something went wrong"
-            }
-        }
-    }
-}
-
-struct MainView: View {
+struct FormView: View {
     @ObservedObject private var viewModel: FormViewModel
 
     init(viewModel: FormViewModel) {
@@ -54,11 +32,9 @@ struct MainView: View {
 }
 
 #Preview {
-    MainView(viewModel: FormViewModel(loader: MockLoader()))
+    FormView(viewModel: FormViewModel(loader: MockLoader()))
 }
 
 class MockLoader: FormLoader {
-    func load(completion: @escaping (FormLoader.Result) -> Void) {
-
-    }
+    func load(completion: @escaping (FormLoader.Result) -> Void) {}
 }
